==> 04_diploma_datasets_mongo_tokens.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Часть 4: запись токенов в базу
1. Считать токены из pickle-файлов
2. Объединить словари
3. Записать в базу
4. Проиндексировать коллекцию

#### Импорт модулей

%load_ext pycodestyle_magic
%pycodestyle_on

import pickle
import gridfs
import pymongo
from tqdm.auto import tqdm
from collections import Counter
from pymongo.errors import DocumentTooLarge, OperationFailure

#### Открытие файлов

with open('presov_tokens.pickle', 'rb') as ft1:
    presov_tokens = pickle.load(ft1)

len(presov_tokens)

with open('sov_tokens.pickle', 'rb') as ft2:
    sov_tokens = pickle.load(ft2)

len(sov_tokens)

with open('postsov_tokens.pickle', 'rb') as ft3:
    postsov_tokens = pickle.load(ft3)

len(postsov_tokens)

#### Объединение

all_tokens = {}

# 1 + 2 и 3
for tok_key, token in tqdm(list(presov_tokens.items())):

    if tok_key not in sov_tokens and tok_key not in postsov_tokens:
        all_tokens[tok_key] = token

    else:
        combined_freq = token['freq']
        combined_freq0 = token['freq_0']
        combined_freq1 = token['freq_1']
        combined_freq2 = token['freq_2']
        combined_docs = token['docs']

        if tok_key in sov_tokens:
            combined_freq += sov_tokens[tok_key]['freq']
            combined_freq0 += sov_tokens[tok_key]['freq_0']
            combined_freq1 += sov_tokens[tok_key]['freq_1']
            combined_freq2 += sov_tokens[tok_key]['freq_2']
            combined_docs.update(sov_tokens[tok_key]['docs'])
            del sov_tokens[tok_key]

        if tok_key in postsov_tokens:
            combined_freq += postsov_tokens[tok_key]['freq']
            combined_freq0 += postsov_tokens[tok_key]['freq_0']
            combined_freq1 += postsov_tokens[tok_key]['freq_1']
            combined_freq2 += postsov_tokens[tok_key]['freq_2']
            combined_docs.update(postsov_tokens[tok_key]['docs'])
            del postsov_tokens[tok_key]

        all_tokens[tok_key] = {'token': token['token'],
                               'lemma': token['lemma'],
                               'pos': token['pos'],
                               'freq': combined_freq,
                               'freq_0': combined_freq0,
                               'freq_1': combined_freq1,
                               'freq_2': combined_freq2,
                               'docs': combined_docs}

# 2 + 3
for tok_key, token in tqdm(list(sov_tokens.items())):

    if tok_key not in postsov_tokens:
        all_tokens[tok_key] = token

    else:
        combined_freq = token['freq'] + postsov_tokens[tok_key]['freq']
        combined_freq0 = token['freq_0'] + postsov_tokens[tok_key]['freq_0']
        combined_freq1 = token['freq_1'] + postsov_tokens[tok_key]['freq_1']
        combined_freq2 = token['freq_2'] + postsov_tokens[tok_key]['freq_2']
        combined_docs = token['docs']
        combined_docs.update(postsov_tokens[tok_key]['docs'])
        del postsov_tokens[tok_key]

        all_tokens[tok_key] = {'token': token['token'],
                               'lemma': token['lemma'],
                               'pos': token['pos'],
                               'freq': combined_freq,
                               'freq_0': combined_freq0,
                               'freq_1': combined_freq1,
                               'freq_2': combined_freq2,
                               'docs': combined_docs}

# то, что есть только в 3
for tok_key, token in tqdm(list(postsov_tokens.items())):
    all_tokens[tok_key] = token

del presov_tokens
del sov_tokens
del postsov_tokens

len(all_tokens)

with open('all_tokens.pickle', 'wb') as ftok:
    pickle.dump(all_tokens, ftok)

#### Считываем токены из файла

with open('all_tokens.pickle', 'rb') as ftok:
    all_tokens = pickle.load(ftok)

# множества не записываются в базу
for tok_key in tqdm(all_tokens):
    all_tokens[tok_key]['docs'] = list(all_tokens[tok_key]['docs'])

**Подключение к базе**

client = pymongo.MongoClient('localhost', 27017)
db = client['thesis']
fs = gridfs.GridFS(db)
db

tokens = db.tokens

# tokens.delete_many({})

#### Запись в базу

def insert_tokens(token_dict, token_collection):
    tokens_to_insert = list(token_dict.values())
    token_collection.insert_many(tokens_to_insert)

# insert_tokens(all_tokens, tokens)  # здесь ошибка DocumentTooLarge

def insert_tokens_one_by_one(token_dict, token_collection):
    big_tokens = []
    for tok_key in tqdm(token_dict):
        try:
            token_collection.insert_one(token_dict[tok_key])
        except (DocumentTooLarge, OperationFailure):
            big_tokens.append(token_dict[tok_key])
    return big_tokens

big_tokens = insert_tokens_one_by_one(all_tokens, tokens)

len(big_tokens)

for i, token in enumerate(tqdm(big_tokens)):
    with open(f'token_{i}.pickle', 'wb') as tok:
        pickle.dump(token, tok)
    with open(f'token_{i}.pickle', 'rb') as tok:
        tok_id = fs.put(tok)
    token['file_id'] = tok_id

for token in tqdm(big_tokens):
    info_to_insert = token
    info_to_insert['docs'] = token['docs'][0:100000]
    tokens.insert_one(info_to_insert)

**Индексация**

tokens.create_index([('token', pymongo.ASCENDING),
                     ('lemma', pymongo.ASCENDING),
                     ('pos', pymongo.ASCENDING)],
                    unique=True)

# tokens.drop_index('token_1_lemma_1_pos_1')
