==> 07_prefixes_frequency_analysis.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Анализ частотности приставок и глаголов
1. В каждом периоде составить список глаголов для каждой приставки
2. Отранжировать глаголы по частотности
3. Сравнить списки по периодам с использованием расстояния Левенштейна и коэффициента корреляции Спирмена

#### Импорт модулей

%load_ext pycodestyle_magic
%pycodestyle_on

import pymongo
import pandas as pd
from tqdm.auto import tqdm
from textdistance import levenshtein
from scipy.stats import spearmanr

#### Получение списка глаголов

pref_df = pd.read_csv('prefixes_and_lemmas.tsv', sep='\t')
pref_df

# list of dicts
pref_dict = pref_df.to_dict(orient='records')

#### Переводим частотности в абсолютные

presov_vol = 72199757
sov_vol = 93031557
postsov_vol = 81392744
total_vol = 246624058

for verb in pref_dict:
    verb['abs_freq'] = verb['freq']
    verb['abs_freq0'] = verb['freq_0']
    verb['abs_freq1'] = verb['freq_1']
    verb['abs_freq2'] = verb['freq_2']
    del verb['freq']
    del verb['freq_0']
    del verb['freq_1']
    del verb['freq_2']
    verb['freq'] = round(verb['abs_freq'] / total_vol * 1000000, 3)
    verb['freq0'] = round(verb['abs_freq0'] / presov_vol * 1000000, 3)
    verb['freq1'] = round(verb['abs_freq1'] / sov_vol * 1000000, 3)
    verb['freq2'] = round(verb['abs_freq2'] / postsov_vol * 1000000, 3)

pref_df_new = pd.DataFrame(pref_dict)
pref_df_new

pref_df_new.to_csv('prefixes_and_lemmas_freq.tsv', sep='\t', index=False)
pref_dict = pref_df_new.to_dict(orient='records')

### Подход №1. Общий список глаголов для всех периодов

prefixes = sorted(list(set([verb['prefix'] for verb in pref_dict])))

prefs_with_verbs = {pref: [verb for verb in pref_dict
                           if verb['prefix'] == pref]
                    for pref in prefixes}

presov_ranged = {pref: [verb['lemma']
                        for verb in sorted(prefs_with_verbs[pref],
                                           key=lambda x: x['freq0'],
                                           reverse=True)]
                 for pref in prefs_with_verbs}

sov_ranged = {pref: [verb['lemma']
                     for verb in sorted(prefs_with_verbs[pref],
                                        key=lambda x: x['freq1'],
                                        reverse=True)]
              for pref in prefs_with_verbs}

postsov_ranged = {pref: [verb['lemma']
                         for verb in sorted(prefs_with_verbs[pref],
                                            key=lambda x: x['freq2'],
                                            reverse=True)]
                  for pref in prefs_with_verbs}

#### Подсчет расстояния Левенштейна и коэффициента корреляции Спирмена

stats = {pref: {'pre_to_sov_lev': levenshtein.distance(presov_ranged[pref],
                                                       sov_ranged[pref]),
                'sov_to_post_lev': levenshtein.distance(sov_ranged[pref],
                                                        postsov_ranged[pref]),
                'pre_to_post_lev': levenshtein.distance(presov_ranged[pref],
                                                        postsov_ranged[pref]),
                'pre_to_sov_sp': spearmanr(presov_ranged[pref],
                                           sov_ranged[pref]),
                'sov_to_post_sp': spearmanr(sov_ranged[pref],
                                            postsov_ranged[pref]),
                'pre_to_post_sp': spearmanr(presov_ranged[pref],
                                            postsov_ranged[pref])}
         for pref in tqdm(prefixes)}

for pref in stats:
    pre_to_sov_sp = str(round(stats[pref]['pre_to_sov_sp'][0], 3))
    pre_to_sov_conf = stats[pref]['pre_to_sov_sp'][1]
    if pre_to_sov_conf < 0.05:
        stats[pref]['pre_to_sov_sp'] = pre_to_sov_sp + '*'
    else:
        stats[pref]['pre_to_sov_sp'] = pre_to_sov_sp

    sov_to_post_sp = str(round(stats[pref]['sov_to_post_sp'][0], 3))
    sov_to_post_conf = stats[pref]['sov_to_post_sp'][1]
    if sov_to_post_conf < 0.05:
        stats[pref]['sov_to_post_sp'] = sov_to_post_sp + '*'
    else:
        stats[pref]['sov_to_post_sp'] = sov_to_post_sp

    pre_to_post_sp = str(round(stats[pref]['pre_to_post_sp'][0], 3))
    pre_to_post_conf = stats[pref]['pre_to_post_sp'][1]
    if pre_to_post_conf < 0.05:
        stats[pref]['pre_to_post_sp'] = pre_to_post_sp + '*'
    else:
        stats[pref]['pre_to_post_sp'] = pre_to_post_sp

pref_stats = pref_df_new.groupby(['prefix']).size().to_dict()

stats_for_df = []
for pref in pref_stats:
    stats_for_df.append({'prefix': pref,
                         'verbs': pref_stats[pref],
                         'pre_to_sov_lev': stats[pref]['pre_to_sov_lev'],
                         'sov_to_post_lev': stats[pref]['sov_to_post_lev'],
                         'pre_to_post_lev': stats[pref]['pre_to_post_lev'],
                         'pre_to_sov_sp': stats[pref]['pre_to_sov_sp'],
                         'sov_to_post_sp': stats[pref]['sov_to_post_sp'],
                         'pre_to_post_sp': stats[pref]['pre_to_post_sp']
                         })

stats_df = pd.DataFrame(stats_for_df)
stats_df

### Подход №2. Разные глаголы для разных периодов, порог частотности = 100
#### Подключение к базе данных

client = pymongo.MongoClient('localhost', 27017)
db = client['thesis']
db

sentences = db.sentences
lemmas = db.lemmas
tokens = db.tokens

#### Извлекаем нужные леммы из базы данных

prefixes = pref_df['prefix'].unique().tolist()

def find_verbs_with_prefix(prefix_name, freq):
    regexp = '^{}'.format(prefix_name)
    found = lemmas.find({'lemma': {'$regex': regexp}, 'pos': 'V',
                         freq: {'$gt': 100}},
                        {'lemma': True, '_id': False}
                        ).sort(freq, -1).limit(100)
    return list(found)

prefix_dict_presov = {prefix: find_verbs_with_prefix(prefix, 'freq_0')
                      for prefix in tqdm(prefixes)}
prefix_dict_sov = {prefix: find_verbs_with_prefix(prefix, 'freq_1')
                   for prefix in tqdm(prefixes)}
prefix_dict_postsov = {prefix: find_verbs_with_prefix(prefix, 'freq_2')
                       for prefix in tqdm(prefixes)}

#### Приводим списки к одинаковой длине

for prefix in prefixes:
    presov = len(prefix_dict_presov[prefix])
    sov = len(prefix_dict_sov[prefix])
    postsov = len(prefix_dict_postsov[prefix])
    min_len = min(presov, sov, postsov)

    if min_len < 100:
        del prefix_dict_presov[prefix]
        del prefix_dict_sov[prefix]
        del prefix_dict_postsov[prefix]
    else:
        prefix_dict_presov[prefix] = [verb['lemma']
                                      for verb in prefix_dict_presov[prefix]]
        prefix_dict_sov[prefix] = [verb['lemma']
                                   for verb in prefix_dict_sov[prefix]]
        prefix_dict_postsov[prefix] = [verb['lemma']
                                       for verb in prefix_dict_postsov[prefix]]

stats = {pref: {'pre_to_sov_lev': levenshtein.distance(prefix_dict_presov[pref],
                                                       prefix_dict_sov[pref]),
                'sov_to_post_lev': levenshtein.distance(prefix_dict_sov[pref],
                                                        prefix_dict_postsov[pref]),
                'pre_to_post_lev': levenshtein.distance(prefix_dict_presov[pref],
                                                        prefix_dict_postsov[pref]),
                'pre_to_sov_sp': spearmanr(prefix_dict_presov[pref],
                                           prefix_dict_sov[pref]),
                'sov_to_post_sp': spearmanr(prefix_dict_sov[pref],
                                            prefix_dict_postsov[pref]),
                'pre_to_post_sp': spearmanr(prefix_dict_presov[pref],
                                            prefix_dict_postsov[pref])}
         for pref in tqdm(list(prefix_dict_presov.keys()))}

for pref in stats:
    pre_to_sov_sp = str(round(stats[pref]['pre_to_sov_sp'][0], 3))
    pre_to_sov_conf = stats[pref]['pre_to_sov_sp'][1]
    if pre_to_sov_conf < 0.05:
        stats[pref]['pre_to_sov_sp'] = pre_to_sov_sp + '*'
    else:
        stats[pref]['pre_to_sov_sp'] = pre_to_sov_sp

    sov_to_post_sp = str(round(stats[pref]['sov_to_post_sp'][0], 3))
    sov_to_post_conf = stats[pref]['sov_to_post_sp'][1]
    if sov_to_post_conf < 0.05:
        stats[pref]['sov_to_post_sp'] = sov_to_post_sp + '*'
    else:
        stats[pref]['sov_to_post_sp'] = sov_to_post_sp

    pre_to_post_sp = str(round(stats[pref]['pre_to_post_sp'][0], 3))
    pre_to_post_conf = stats[pref]['pre_to_post_sp'][1]
    if pre_to_post_conf < 0.05:
        stats[pref]['pre_to_post_sp'] = pre_to_post_sp + '*'
    else:
        stats[pref]['pre_to_post_sp'] = pre_to_post_sp

stats_for_df = []
for pref in list(prefix_dict_presov.keys()):
    stats_for_df.append({'prefix': pref,
                         'verbs': len(prefix_dict_presov[pref]),
                         'pre_to_sov_lev': stats[pref]['pre_to_sov_lev'],
                         'sov_to_post_lev': stats[pref]['sov_to_post_lev'],
                         'pre_to_post_lev': stats[pref]['pre_to_post_lev'],
                         'pre_to_sov_sp': stats[pref]['pre_to_sov_sp'],
                         'sov_to_post_sp': stats[pref]['sov_to_post_sp'],
                         'pre_to_post_sp': stats[pref]['pre_to_post_sp']
                         })

stats_df = pd.DataFrame(stats_for_df)
stats_df
