==> 09_grammatical_profiles_tokens.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Часть 9: получение грамматических профилей глаголов
1. Выбрать из базы предложения нужных периодов
2. Для каждого исследуемого глагола получить набор токенов с грамматическими характеристиками
3. Сохранить в pickle-файлы

#### Импорт модулей

%load_ext pycodestyle_magic
%pycodestyle_on

import pickle
import pymongo
import pandas as pd
from tqdm.auto import tqdm
from collections import Counter

**Подключение к базе**

client = pymongo.MongoClient('localhost', 27017)
db = client['thesis']
# fs = gridfs.GridFS(db)
db

sentences = db.sentences
lemmas = db.lemmas
tokens = db.tokens

#### Получение списка глаголов

pref_df = pd.read_csv('prefixes_and_lemmas.tsv', sep='\t')
pref_df

# list of dicts
pref_dict = pref_df.to_dict(orient='records')

verbs_to_search = sorted(set([v['lemma'] for v in pref_dict]))

#### Поиск глагольных употреблений в базе

# досоветский период

presov_verb_dict = {verb: [] for verb in verbs_to_search}

for sentence in tqdm(sentences.find({'period': 0},
                                    {'token_info': True, '_id': False})):
    verb_list = [token for token in sentence['token_info']
                 if token['pos'] == 'V']
    for verb in verb_list:
        if verb['lemma'] in verbs_to_search:
            presov_verb_dict[verb['lemma']].append({'token': verb['token'],
                                                    'gram': verb['gram'].split('|')[0]})

with open('presov_verb_dict.pickle', 'wb') as f:
    pickle.dump(presov_verb_dict, f)

# советский период

sov_verb_dict = {verb: [] for verb in verbs_to_search}

for sentence in tqdm(sentences.find({'period': 1},
                                    {'token_info': True, '_id': False})):
    verb_list = [token for token in sentence['token_info']
                 if token['pos'] == 'V']
    for verb in verb_list:
        if verb['lemma'] in verbs_to_search:
            sov_verb_dict[verb['lemma']].append({'token': verb['token'],
                                                 'gram': verb['gram'].split('|')[0]})

with open('sov_verb_dict.pickle', 'wb') as f:
    pickle.dump(sov_verb_dict, f)

# постсоветский период

postsov_verb_dict = {verb: [] for verb in verbs_to_search}

for sentence in tqdm(sentences.find({'period': 2},
                                    {'token_info': True, '_id': False})):
    verb_list = [token for token in sentence['token_info']
                 if token['pos'] == 'V']
    for verb in verb_list:
        if verb['lemma'] in verbs_to_search:
            postsov_verb_dict[verb['lemma']].append({'token': verb['token'],
                                                     'gram': verb['gram'].split('|')[0]})

with open('postsov_verb_dict.pickle', 'wb') as f:
    pickle.dump(postsov_verb_dict, f)
