==> 11_fasttext_training.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Часть 11: обучение fasttext-моделей

#### Импорт модулей

%load_ext pycodestyle_magic
%pycodestyle_on

import fasttext

#### Обучение моделей
На тех же данных и с теми же параметрами, что и базовый word2vec (т.е. леммы без пунктуации). Алгоритм -- CBOW.

# досоветский период
presov_ft_model = fasttext.train_unsupervised('word2vec_presov.txt',
                                              model='cbow')

len(presov_ft_model.words)

presov_ft_model.save_model('fasttext_presov_lemmas.bin')

# советский период
sov_ft_model = fasttext.train_unsupervised('word2vec_sov.txt',
                                           model='cbow')

len(sov_ft_model.words)

sov_ft_model.save_model('fasttext_sov_lemmas.bin')

# постсоветский период
postsov_ft_model = fasttext.train_unsupervised('word2vec_postsov.txt',
                                               model='cbow')

len(postsov_ft_model.words)

postsov_ft_model.save_model('fasttext_postsov_lemmas.bin')
