==> 12_grammatical_profiles_getting_vectors.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Получение и анализ грамматических профилей глаголов
1. Считать подготовленные pickle-файлы с грамматической информацией о токенах
2. Составить представления глаголов на основе их токенов

#### Импорт модулей

# %load_ext pycodestyle_magic
# %pycodestyle_on

import re
import pickle
import numpy as np
import pandas as pd
from tqdm.auto import tqdm
from collections import Counter
from statistics import mean
from scipy.spatial.distance import cityblock

#### Получение списка глаголов

pref_df = pd.read_csv('prefixes_and_lemmas_dropna.tsv', sep='\t')
pref_df

# list of dicts
pref_dict = pref_df.to_dict(orient='records')

verbs_to_search = sorted(set([v['lemma'] for v in pref_dict]))

#### Получение грамматических профилей
Открываем ранее полученные файлы:

with open('presov_verb_dict.pickle', 'rb') as f:
    presov_verb_dict = pickle.load(f)

with open('sov_verb_dict.pickle', 'rb') as f:
    sov_verb_dict = pickle.load(f)

with open('postsov_verb_dict.pickle', 'rb') as f:
    postsov_verb_dict = pickle.load(f)

(Giulianelli et al. 2022):
- Tense : {Past 42, Pres 51}
- VerbForm : {Part 68, Fin 25, Inf 9}
- Mood : {Ind 25}
- Voice : {Pass 17}

Для нашего датасета:
- tense: pres, inpr, past
- verbform: ger, part, inf, fin
- mood: ind, imp
- number: sg, pl
- person: 1p, 2p, 3p
- gender: m, f, n
- voice: act, pass

Другое:
- вид: perf, imperf
- переходность: trans, intrans
- падеж: есть только у причастий
- форма: краткая/полная -- есть только у причастий
- одушевленность: вообще что-то странное

# full = {'tense': {'непрош': 0, 'прош': 0},
#         'vform': {'прич': 0, 'деепр': 0, 'инф': 0, 'фин': 0},
#         'mood': {'пов': 0, 'инд': 0},
#         'numb': {'ед': 0, 'мн': 0},
#         'pers': {'1-л': 0, '2-л': 0, '3-л': 0},
#         'gend': {'муж': 0, 'жен': 0, 'сред': 0},
#         'voice': {'действ': 0, 'страд': 0}
#         }

# presov_gram_dict = {}
# gram_list = set(['непрош', 'прош', 'прич', 'деепр', 'инф', 'фин', 'пов', 'изъяв', 'ед', 'мн', '1-л', '2-л', '3-л', 'муж', 'жен', 'сред', 'действ', 'страд'])
# # grams_to_ignore = set(['полн', 'кр', 'им', 'вин', 'дат', 'твор', 'пр', 'од', 'неод'])

# for verb in tqdm(presov_verb_dict):
#     presov_gram_dict[verb] = {'непрош': 0, 'прош': 0,
#                               'прич': 0, 'деепр': 0, 'инф': 0, 'фин': 0,
#                               'пов': 0, 'изъяв': 0,
#                               'ед': 0, 'мн': 0,
#                               '1-л': 0, '2-л': 0, '3-л': 0,
#                               'муж': 0, 'жен': 0, 'сред': 0,
#                               'действ': 0, 'страд': 0}
#     for token in presov_verb_dict[verb]:
#         gram_info = re.sub('^V[=,]', '', token['gram'])
#         # убрать одушевленность
#         gram_info = re.sub('сов|несов|нп|пе|устар|обсц|разг|полн|кр|им|вин|род|дат|твор|од|неод', '', gram_info)
#         gram_info = re.sub(',+', ',', gram_info).strip('(,=').split(',')
#         for gram in gram_info:
#             if gram in gram_list:
#                 presov_gram_dict[verb][gram] += 1
#         # не нужно ли это убрать?
#         if 'прич' not in gram_info and 'деепр' not in gram_info and 'инф' not in gram_info:
#             presov_gram_dict[verb]['фин'] += 1

def get_gram_profiles(verb_dict, all_grams=set()):

    gram_dict = {}

    for verb in tqdm(verb_dict):
        gram_list = []

        for token in verb_dict[verb]:
            gram_info = re.sub('^V[=,]', '', token['gram'])
            gram_info = re.sub('сов|несов|нп|пе|устар|обсц|разг|полн|кр|им|вин|род|дат|твор|од|неод',
                               '', gram_info)
            gram_info = re.sub(',+', ',', gram_info).strip('(,=').split(',')
            gram_info = sorted(set([gram for gram in gram_info if gram != 'пр']))
            gram = '_'.join(gram_info)
            gram_list.append(gram)
            all_grams.add(gram)

        gram_counter = list(Counter(gram_list).items())

        # нормализация
        tok_num = len(verb_dict[verb])
        for i, gram in enumerate(gram_counter):
            norm_gram = gram[1] / tok_num
            gram_counter[i] = (gram[0], norm_gram)

        # сохраняем грамматический профиль глагола
        gram_dict[verb] = gram_counter

    return gram_dict, all_grams

presov_gram_dict, all_grams = get_gram_profiles(presov_verb_dict)

presov_gram_dict['беспокоиться']

sov_gram_dict, all_grams = get_gram_profiles(sov_verb_dict, all_grams)

postsov_gram_dict, all_grams = get_gram_profiles(postsov_verb_dict, all_grams)

#### Получаем векторы глаголов
Сначала нужно убедиться, что словарь с каждым профилем содержит все возможные ключи.

def unify_gram_keys(gram_dict, all_grams):

    for verb in gram_dict:
        verb_forms = set([form[0] for form in gram_dict[verb]])
        missing_grams = all_grams - verb_forms
        for gram in missing_grams:
            gram_dict[verb].append((gram, 0))
        gram_dict[verb] = sorted(gram_dict[verb], key=lambda x: x[0])

    return gram_dict

presov_gram_dict = unify_gram_keys(presov_gram_dict, all_grams)
sov_gram_dict = unify_gram_keys(sov_gram_dict, all_grams)
postsov_gram_dict = unify_gram_keys(postsov_gram_dict, all_grams)

Теперь переводим словари в векторы.

presov_vect_dict = {verb: np.array([gram[1] for gram in presov_gram_dict[verb]])
                    for verb in presov_gram_dict}

len(presov_verb_dict['беспокоиться'])

presov_vect_dict['беспокоиться']

sov_vect_dict = {verb: np.array([gram[1] for gram in sov_gram_dict[verb]])
                 for verb in sov_gram_dict}
postsov_vect_dict = {verb: np.array([gram[1] for gram in postsov_gram_dict[verb]])
                     for verb in postsov_gram_dict}

for verb in pref_dict:
    lemma = verb['lemma']
    verb['gram_presov'] = presov_vect_dict[lemma]
    verb['gram_sov'] = sov_vect_dict[lemma]
    verb['gram_postsov'] = postsov_vect_dict[lemma]

#### Рассчитываем степень изменения между периодами

for verb in tqdm(pref_dict):
    verb['gram_pre_to_sov'] = round(cityblock(verb['gram_presov'], verb['gram_sov']), 3)
    verb['gram_sov_to_post'] = round(cityblock(verb['gram_sov'], verb['gram_postsov']), 3)
    verb['gram_pre_to_post'] = round(cityblock(verb['gram_presov'], verb['gram_postsov']), 3)

#### Ранжируем по степени изменения в каждом периоде

verbs_change_pre_to_sov = [{'lemma': verb['lemma'],
                            'abs_freq0': verb['abs_freq0'],
                            'abs_freq1': verb['abs_freq1'],
                            'abs_freq2': verb['abs_freq2']}
                           for verb in sorted(pref_dict,
                                              key=lambda x:x['gram_pre_to_sov'],
                                              reverse=True)]
verbs_change_sov_to_post = [{'lemma': verb['lemma'],
                            'abs_freq0': verb['abs_freq0'],
                            'abs_freq1': verb['abs_freq1'],
                            'abs_freq2': verb['abs_freq2']}
                           for verb in sorted(pref_dict,
                                              key=lambda x:x['gram_sov_to_post'],
                                              reverse=True)]
verbs_change_pre_to_post = [{'lemma': verb['lemma'],
                            'abs_freq0': verb['abs_freq0'],
                            'abs_freq1': verb['abs_freq1'],
                            'abs_freq2': verb['abs_freq2']}
                           for verb in sorted(pref_dict,
                                              key=lambda x:x['gram_pre_to_post'],
                                              reverse=True)]

verbs_change_pre_to_sov = [verb['lemma'] for verb in verbs_change_pre_to_sov
                           if verb['abs_freq0'] > 100 and verb['abs_freq1'] > 100 and verb['abs_freq2'] > 100][0:200]
verbs_change_sov_to_post = [verb['lemma'] for verb in verbs_change_sov_to_post
                            if verb['abs_freq0'] > 100 and verb['abs_freq1'] > 100 and verb['abs_freq2'] > 100][0:200]
verbs_change_pre_to_post = [verb['lemma'] for verb in verbs_change_pre_to_post
                            if verb['abs_freq0'] > 100 and verb['abs_freq1'] > 100 and verb['abs_freq2'] > 100][0:200]

#### Выбираем топ-10 наиболее изменившихся глаголов для дальнейшего анализа в ELMo

verbs_for_elmo = []
for verb in verbs_change_pre_to_sov:
    if verb in verbs_change_sov_to_post and verb in verbs_change_pre_to_post:
        verbs_for_elmo.append(verb)
    if len(verbs_for_elmo) == 10:
        break

print(verbs_for_elmo)

#### Рассчитываем изменения в приставках

prefixes = sorted(list(set([verb['prefix'] for verb in pref_dict])))

gram_tok_dict = {prefix: [] for prefix in prefixes}

for verb in pref_dict:
    gram_tok_dict[verb['prefix']].append(verb)

for prefix in gram_tok_dict:
    verbs = gram_tok_dict[prefix]
    pre_to_sov = [(verb['lemma'], verb['gram_pre_to_sov'])
                  for verb in sorted(verbs,
                                     key=lambda x: x['gram_pre_to_sov'],
                                     reverse=True)]
    sov_to_post = [(verb['lemma'], verb['gram_sov_to_post'])
                   for verb in sorted(verbs,
                                      key=lambda x: x['gram_sov_to_post'],
                                      reverse=True)]
    pre_to_post = [(verb['lemma'], verb['gram_pre_to_post'])
                   for verb in sorted(verbs,
                                      key=lambda x: x['gram_pre_to_post'],
                                      reverse=True)]
    gram_tok_dict[prefix] = {'pre_to_sov_mean': round(mean([v[1] for v in pre_to_sov]), 3),
                             'sov_to_post_mean': round(mean([v[1] for v in sov_to_post]), 3),
                             'pre_to_post_mean': round(mean([v[1] for v in pre_to_post]), 3),
                             'pre_to_sov_ranged': pre_to_sov,
                             'sov_to_post_ranged': sov_to_post,
                             'pre_to_post_ranged': pre_to_post}

gram_tok_dict['полу']['sov_to_post_mean']

#### Ранжируем приставки по степени изменения

prefs_pre_to_sov_ranged = [(pref,
                            len(gram_tok_dict[pref]['pre_to_sov_ranged']),
                            gram_tok_dict[pref]['pre_to_sov_mean'])
                           for pref in sorted(gram_tok_dict,
                                              key=lambda x: gram_tok_dict[x]['pre_to_sov_mean'],
                                              reverse=True)]

prefs_sov_to_post_ranged = [(pref,
                             len(gram_tok_dict[pref]['sov_to_post_ranged']),
                             gram_tok_dict[pref]['sov_to_post_mean'])
                            for pref in sorted(gram_tok_dict,
                                               key=lambda x: gram_tok_dict[x]['sov_to_post_mean'],
                                               reverse=True)]

prefs_pre_to_post_ranged = [(pref,
                             len(gram_tok_dict[pref]['pre_to_post_ranged']),
                             gram_tok_dict[pref]['pre_to_post_mean'])
                            for pref in sorted(gram_tok_dict,
                                               key=lambda x: gram_tok_dict[x]['pre_to_post_mean'],
                                               reverse=True)]

pref_ranged_df = pd.DataFrame(
    {'gram_pre_to_sov': prefs_pre_to_sov_ranged,
     'gram_sov_to_post': prefs_sov_to_post_ranged,
     'gram_pre_to_post': prefs_pre_to_post_ranged,
     })

pref_ranged_df

ranged_dict = pref_ranged_df.to_dict(orient='records')

# приводим к стандартному виду
convt = {'бе[зс]': 'без-',
         'в[зс]': 'вз-',
         'во?': 'в-',
         'во[зс]': 'воз-',
         'вы': 'вы-',
         'до': 'до-',
         'за': 'за-',
         'и[зс]': 'из-',
         'на': 'на-',
         'над': 'над-',
         'недо': 'недо-',
         'ни[зс]': 'низ-',
         'о': 'о-',
         'обе[зс]': 'обез-',
         'от': 'от-',
         'пере': 'пере-',
         'по': 'по-',
         'под': 'под-',
         'полу': 'полу-',
         'пре': 'пре-',
         'пред': 'пред-',
         'при': 'при-',
         'про': 'про-',
         'ра[зс]': 'раз-',
         'с': 'с-',
         'у': 'у-'}

for row in ranged_dict:
    for col in row:
        row[col] = convt[row[col][0]]

clean_ranged_df = pd.DataFrame(ranged_dict)
clean_ranged_df
