==> 16_elmo_sentence_extraction_for_clustering.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Получение предложений с наиболее изменившимся глаголами для кластеризации ELMo

#### Импорт модулей

# %load_ext pycodestyle_magic
# %pycodestyle_on

import random
import pickle
import pymongo
import pandas as pd
from tqdm.auto import tqdm

random.seed(42)

#### Подключение к базе данных

client = pymongo.MongoClient('localhost', 27017)
db = client['thesis']
db

sentences = db.sentences
lemmas = db.lemmas
tokens = db.tokens

#### Получение списка глаголов

verbs_to_search = ['заметать', 'постановлять', 'изготавливаться', 'наследовать', 'претерпеть', 'взводить', 'засылать', 'дозволять', 'устрашать', 'округлять']

#### Поиск предложений с глаголами в базе

# для каждого глагола получаем список предложений, в которых он встречается
verbs_with_sents = {}
for verb in tqdm(verbs_to_search):
    sents_with_verb = lemmas.find_one({'lemma': verb, 'pos': 'V'},
                                      {'docs': True, '_id': False})
    verbs_with_sents[verb] = sents_with_verb['docs']

presov_sents = {verb: [] for verb in verbs_with_sents}
sov_sents = {verb: [] for verb in verbs_with_sents}
postsov_sents = {verb: [] for verb in verbs_with_sents}

ids_to_search = {verb: random.sample(verbs_with_sents[verb],
                                     min(len(verbs_with_sents[verb]), 500))
                 for verb in verbs_with_sents}

for verb in tqdm(verbs_with_sents):
    for sentence in sentences.find({'_id': {'$in': ids_to_search[verb]}, 'period': 0},
                                   {'token_info': True, '_id': True}).limit(50):
        tok_list = [tok['token'] for tok in sentence['token_info']]
        target_tok = [tok['lemma'] for tok in sentence['token_info']].index(verb)
        presov_sents[verb].append((tok_list, target_tok))
        ids_to_search[verb].remove(sentence['_id'])

with open('presov_sents_for_clust.pickle', 'wb') as f:
    pickle.dump(presov_sents, f)

del presov_sents

for verb in tqdm(verbs_with_sents):
    for sentence in sentences.find({'_id': {'$in': ids_to_search[verb]}, 'period': 1},
                                   {'token_info': True, '_id': True}).limit(50):
        tok_list = [tok['token'] for tok in sentence['token_info']]
        target_tok = [tok['lemma'] for tok in sentence['token_info']].index(verb)
        sov_sents[verb].append((tok_list, target_tok))
        ids_to_search[verb].remove(sentence['_id'])

with open('sov_sents_for_clust.pickle', 'wb') as f:
    pickle.dump(sov_sents, f)

del sov_sents

for verb in tqdm(verbs_with_sents):
    for sentence in sentences.find({'_id': {'$in': ids_to_search[verb]}, 'period': 2},
                                   {'token_info': True, '_id': True}).limit(50):
        tok_list = [tok['token'] for tok in sentence['token_info']]
        target_tok = [tok['lemma'] for tok in sentence['token_info']].index(verb)
        postsov_sents[verb].append((tok_list, target_tok))
        ids_to_search[verb].remove(sentence['_id'])

with open('postsov_sents_for_clust.pickle', 'wb') as f:
    pickle.dump(postsov_sents, f)

del postsov_sents
