==> 17_elmo_get_embeddings_for_clustering.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Получение эмбеддингов ELMo для наиболее изменившихся глаголов

#### Импорт модулей

# %load_ext pycodestyle_magic
# %pycodestyle_on

!pip install simple_elmo

import pickle
import random
import pandas as pd
from tqdm.auto import tqdm
from simple_elmo import ElmoModel

import warnings
warnings.filterwarnings('ignore')

#### Получение списка глаголов и предложений

with open('../input/elmornc3/presov_sents_for_clust.pickle', 'rb') as f:
    presov_sents = pickle.load(f)

with open('../input/elmornc3/sov_sents_for_clust.pickle', 'rb') as f:
    sov_sents = pickle.load(f)

with open('../input/elmornc3/postsov_sents_for_clust.pickle', 'rb') as f:
    postsov_sents = pickle.load(f)

verbs_to_search = ['заметать', 'постановлять', 'изготавливаться', 'наследовать', 'претерпеть', 'взводить', 'засылать', 'дозволять', 'устрашать', 'округлять']

#### Загрузка модели

model = ElmoModel()
model.load('../input/elmo-pretrained/ruwikiruscorpora_tokens_elmo_1024_2019',
           max_batch_size=50)

#### Получение эмбеддингов ELMo
(Ryzhova et al. 2021)

For each word we find all the sentences where these words occur and construct 100 random pairs. The first sentence in a pair belongs to one time period, the second to the other period. For each sentence we compute the ELMo embedding of the word for which we want to measure the semantic shift value. The final score is the average cosine similarity for all pairs. To achieve a certain level of robustness, we take the average value of the cosine similarity from 5 experiments on different 100 random pairs in each iteration.

for idx, verb in enumerate(verbs_to_search):

    presov_inputs = [sent[0] for sent in presov_sents[verb]]
    presov_embeds = model.get_elmo_vectors(presov_inputs)
    for i, sent in enumerate(presov_sents[verb]):
        tup_to_list = list(sent)
        target_emb = presov_embeds[i][tup_to_list[1]]
        tup_to_list.append(target_emb)
        presov_sents[verb][i] = tup_to_list

    print(f'Completed run {idx+1} of {len(verbs_to_search)}')

with open('presov_samples_clust.pickle', 'wb') as f:
    pickle.dump(presov_sents, f)

for idx, verb in enumerate(verbs_to_search):

    sov_inputs = [sent[0] for sent in sov_sents[verb]]
    sov_embeds = model.get_elmo_vectors(sov_inputs)
    for i, sent in enumerate(sov_sents[verb]):
        tup_to_list = list(sent)
        target_emb = sov_embeds[i][tup_to_list[1]]
        tup_to_list.append(target_emb)
        sov_sents[verb][i] = tup_to_list
    
    print(f'Completed run {idx+1} of {len(verbs_to_search)}')

with open('sov_samples_clust.pickle', 'wb') as f:
    pickle.dump(sov_sents, f)

for idx, verb in enumerate(verbs_to_search):

    postsov_inputs = [sent[0] for sent in postsov_sents[verb]]
    postsov_embeds = model.get_elmo_vectors(postsov_inputs)
    for i, sent in enumerate(postsov_sents[verb]):
        tup_to_list = list(sent)
        target_emb = postsov_embeds[i][tup_to_list[1]]
        tup_to_list.append(target_emb)
        postsov_sents[verb][i] = tup_to_list
    
    print(f'Completed run {idx+1} of {len(verbs_to_search)}')

with open('postsov_samples_clust.pickle', 'wb') as f:
    pickle.dump(postsov_sents, f)
