==> 19_results_spearman_correlation.ipynb <==
# Микродиахроническое исследование русских приставок методами дистрибутивной семантики
## Автор: Елизавета Клыкова, БКЛ181
### Подсчет корреляции результатов разных моделей
С помощью коэффициента корреляции Спирмена

#### Импорт модулей

# %load_ext pycodestyle_magic
# %pycodestyle_on

import pandas as pd
from scipy.stats import spearmanr

#### Получение списка глаголов

rank_df = pd.read_csv('prefixes_ranked.tsv', sep='\t')
rank_df

pre_to_sov_df = rank_df[list(rank_df.columns)[0:6]]
for col in list(rank_df.columns)[0:6]:
    pre_to_sov_df[col.split('_')[0]] = pre_to_sov_df[col].astype('category').cat.codes
pre_to_sov_df.drop(columns=list(rank_df.columns)[0:6], inplace=True)

pre_to_sov_df.corr(method='spearman')

sov_to_post_df = rank_df[list(rank_df.columns)[6:12]]
for col in list(rank_df.columns)[6:12]:
    sov_to_post_df[col.split('_')[0]] = sov_to_post_df[col].astype('category').cat.codes
sov_to_post_df.drop(columns=list(rank_df.columns)[6:12], inplace=True)

pre_to_post_df = rank_df[list(rank_df.columns)[12:18]]
for col in list(rank_df.columns)[12:18]:
    pre_to_post_df[col.split('_')[0]] = pre_to_post_df[col].astype('category').cat.codes
pre_to_post_df.drop(columns=list(rank_df.columns)[12:18], inplace=True)

((pre_to_sov_df.corr(method='spearman') + sov_to_post_df.corr(method='spearman') + pre_to_post_df.corr(method='spearman')) / 3).round(3)
